Code Refactoring: Created Infra
