+++ garage/src/main/java/ai/smartdoc/garage/chat/internal/service/ChatPortAdapter.java
@@ -3,7 +3,7 @@ package ai.smartdoc.garage.chat.internal.service;
 import ai.smartdoc.garage.chat.ChatPort;
 import ai.smartdoc.garage.chat.internal.entity.Chat;
 import ai.smartdoc.garage.chat.internal.entity.Message;
-import ai.smartdoc.garage.common.dto.UploadResponse;
+import ai.smartdoc.garage.chat.internal.dto.UploadResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
