+++ garage/src/main/java/ai/smartdoc/garage/chat/internal/service/FileService.java
@@ -3,10 +3,10 @@ package ai.smartdoc.garage.chat.internal.service;
 import ai.smartdoc.garage.chat.internal.dao.FileDao;
 import ai.smartdoc.garage.chat.internal.entity.Chat;
 import ai.smartdoc.garage.chat.internal.entity.Chunk;
-import ai.smartdoc.garage.common.dto.UploadResponse;
+import ai.smartdoc.garage.chat.internal.dto.UploadResponse;
 import ai.smartdoc.garage.common.exception.GarageException;
-import ai.smartdoc.garage.huggingface.HuggingFacePort;
-import ai.smartdoc.garage.qdrant.QdrantPort;
+import ai.smartdoc.garage.infra.huggingface.HuggingFacePort;
+import ai.smartdoc.garage.infra.qdrant.QdrantPort;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.text.PDFTextStripper;
 import org.apache.poi.xwpf.usermodel.XWPFDocument;
