+++ garage/src/main/java/ai/smartdoc/garage/chat/internal/service/MessageService.java
@@ -6,12 +6,12 @@ import ai.smartdoc.garage.chat.internal.dao.MessageDao;
 import ai.smartdoc.garage.chat.internal.entity.Chat;
 import ai.smartdoc.garage.chat.internal.entity.Chunk;
 import ai.smartdoc.garage.chat.internal.entity.Message;
-import ai.smartdoc.garage.cohere.CoherePort;
+import ai.smartdoc.garage.infra.cohere.CoherePort;
 import ai.smartdoc.garage.common.dto.CohereRerankResponse;
 import ai.smartdoc.garage.common.exception.GarageException;
 import ai.smartdoc.garage.common.utils.IdCreator;
-import ai.smartdoc.garage.huggingface.HuggingFacePort;
-import ai.smartdoc.garage.qdrant.QdrantPort;
+import ai.smartdoc.garage.infra.huggingface.HuggingFacePort;
+import ai.smartdoc.garage.infra.qdrant.QdrantPort;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
