+++ garage/src/main/java/ai/smartdoc/garage/infra/qdrant/internal/QdrantPoint.java
@@ -1,4 +1,4 @@
-package ai.smartdoc.garage.qdrant.internal;
+package ai.smartdoc.garage.infra.qdrant.internal;
 
 import lombok.Data;
 import java.util.List;
